field_type* changed to unsigned short

+++ include/dbi/dbi.h
@@ -18,7 +18,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  * 
- * $Id: dbi.h,v 1.10 2001/04/08 21:05:20 mmt Exp $
+ * $Id: dbi.h,v 1.11 2001/04/11 03:51:15 mmt Exp $
  */
 
 #ifndef __DBI_H__
@@ -45,8 +45,8 @@ typedef struct dbi_row_s {
 	void *row_handle; /* will be typecast into driver-specific type */
 	unsigned int numfields;
 	char **field_names;
-	int *field_types;
-	unsigned int *field_type_attributes;
+	unsigned short *field_types;
+	unsigned short *field_type_attributes;
 	void **field_values;
 	struct dbi_row_s *next; /* NULL, unless we slurp all available rows at once */
 } dbi_row_t;
